Attempt to support resumeSeries method with HBO Go
I'm not convinced the receiver app actually updates the marker
automatically, but... maybe. Incorporating that into this library would
be tricky for CLI use (since, unlike the other services, we'd have to
keep the app running after the show starts in order to keep that
updated) but in a long-running application it might be reasonable....

Will test later to see if it's even a problem
